+++ src/meltano/cli/install.py
@@ -1,8 +1,8 @@
 import click
 
+from meltano.core.legacy_tracking import LegacyTracker
 from meltano.core.plugin import PluginType
 from meltano.core.project_plugins_service import ProjectPluginsService
-from meltano.core.tracking import GoogleAnalyticsTracker
 
 from . import cli
 from .params import pass_project
@@ -47,7 +47,7 @@ def install(project, plugin_type, plugin_name, clean, parallelism):
 
     success = install_plugins(project, plugins, parallelism=parallelism, clean=clean)
 
-    tracker = GoogleAnalyticsTracker(project)
+    tracker = LegacyTracker(project)
     tracker.track_meltano_install()
 
     if not success: