+++ src/meltano/cli/ui.py
@@ -8,12 +8,12 @@ import signal
 import click
 from click_default_group import DefaultGroup
 
-from meltano.api.workers import APIWorker, MeltanoCompilerWorker, UIAvailableWorker
+from meltano.api.workers import APIWorker, UIAvailableWorker
+from meltano.core.legacy_tracking import LegacyTracker
 from meltano.core.project_settings_service import (
     ProjectSettingsService,
     SettingValueStore,
 )
-from meltano.core.tracking import GoogleAnalyticsTracker
 
 from . import cli
 from .params import pass_project
@@ -104,19 +104,12 @@ def start(ctx, reload, bind, bind_port):
         ProjectSettingsService.config_override["ui.bind_port"] = bind_port
 
     project = ctx.obj["project"]
-    tracker = GoogleAnalyticsTracker(project)
+    tracker = LegacyTracker(project)
     tracker.track_meltano_ui()
 
     ensure_secure_setup(project)
 
     workers = []
-
-    try:
-        compiler_worker = MeltanoCompilerWorker(project)
-        compiler_worker.compiler.compile()
-        workers.append(compiler_worker)
-    except Exception as exn:
-        logger.error(f"Initial compilation failed: {exn}")
 
     workers.append(UIAvailableWorker(project))
     workers.append(