+++ src/meltano/cli/utils.py
@@ -10,8 +10,8 @@ import click
 
 from meltano.core.logging import setup_logging
 from meltano.core.plugin import PluginType
-from meltano.core.plugin.base import PluginRef
 from meltano.core.plugin.error import PluginNotFoundError
+from meltano.core.plugin.project_plugin import ProjectPlugin
 from meltano.core.plugin_install_service import (
     PluginInstallReason,
     PluginInstallService,
@@ -20,6 +20,7 @@ from meltano.core.plugin_install_service import (
 from meltano.core.plugin_lock_service import LockfileAlreadyExistsError
 from meltano.core.project import Project
 from meltano.core.project_add_service import (
+    PluginAddedReason,
     PluginAlreadyAddedException,
     ProjectAddService,
 )
@@ -51,11 +52,16 @@ class CliError(Exception):
         self.printed = True
 
 
-def print_added_plugin(plugin, related=False):
+def print_added_plugin(
+    plugin,
+    reason: PluginAddedReason = PluginAddedReason.USER_REQUEST,
+):
     """Print added plugin."""
     descriptor = plugin.type.descriptor
-    if related:
+    if reason is PluginAddedReason.RELATED:
         descriptor = f"related {descriptor}"
+    elif reason is PluginAddedReason.REQUIRED:
+        descriptor = f"required {descriptor}"
 
     if plugin.should_add_to_file():
         click.secho(
@@ -374,8 +380,8 @@ def add_plugin(
 
 
 def add_related_plugins(
-    project,
-    plugins,
+    project: Project,
+    plugins: List[ProjectPlugin],
     add_service: ProjectAddService,
     plugin_types: List[PluginType],
 ):
@@ -386,10 +392,32 @@ def add_related_plugins(
             plugin_install, plugin_types=plugin_types
         )
         for related_plugin in related_plugins:
-            print_added_plugin(related_plugin, related=True)
+            print_added_plugin(related_plugin, reason=PluginAddedReason.RELATED)
             click.echo()
 
         added_plugins.extend(related_plugins)
+
+    return added_plugins
+
+
+def add_required_plugins(
+    project: Project,
+    plugins: List[ProjectPlugin],
+    add_service: ProjectAddService,
+    plugin_types: List[PluginType] = None,
+):
+    """Add any Plugins required by the given Plugin."""
+    plugin_types = plugin_types or list(PluginType)
+    added_plugins = []
+    for plugin_install in plugins:
+        required_plugins = add_service.add_related(
+            plugin_install, plugin_types=plugin_types
+        )
+        for required_plugin in required_plugins:
+            print_added_plugin(required_plugin, reason=PluginAddedReason.REQUIRED)
+            click.echo()
+
+        added_plugins.extend(required_plugins)
 
     return added_plugins
 
@@ -470,7 +498,9 @@ def propagate_stop_signals(proc):
         signal.signal(signal.SIGTERM, original_termination_handler)
 
 
-def check_dependencies_met(plugins, plugins_service: ProjectPluginsService) -> Tuple[bool, str]:
+def check_dependencies_met(
+    plugins, plugins_service: ProjectPluginsService
+) -> Tuple[bool, str]:
     """Check dependencies of added plugins are met.
 
     Args: