+++ src/meltano/core/plugin_lock_service.py
@@ -8,8 +8,6 @@ from pathlib import Path
 from structlog.stdlib import get_logger
 
 from meltano.core.plugin.base import BasePlugin, PluginRef, StandalonePlugin, Variant
-from meltano.core.plugin.project_plugin import ProjectPlugin
-from meltano.core.plugin_discovery_service import PluginDiscoveryService
 from meltano.core.project import Project
 
 logger = get_logger(__name__)
@@ -34,19 +32,17 @@ class LockfileAlreadyExistsError(Exception):
 class PluginLockService:
     """Plugin Lockfile Service."""
 
-    def __init__(self, project: Project, discovery_service: PluginDiscoveryService):
+    def __init__(self, project: Project):
         """Create a new Plugin Lockfile Service.
 
         Args:
             project: The Meltano project.
-            discovery_service: The plugin discovery service.
         """
         self.projet = project
-        self.discovery_service = discovery_service
 
     def save(
         self,
-        plugin: ProjectPlugin | BasePlugin,
+        plugin: BasePlugin,
         *,
         overwrite: bool = False,
         exists_ok: bool = False,
@@ -66,28 +62,12 @@ class PluginLockService:
 
         logger.info(f"Locking a {type(plugin)}")
 
-        if isinstance(plugin, BasePlugin):
-            plugin_def = plugin.definition
-            path = self.projet.plugin_lock_path(
-                plugin_def.type,
-                plugin_def.name,
-                variant_name=variant,
-            )
-
-        elif plugin.inherit_from is None:
-            path = self.projet.plugin_lock_path(
-                plugin.type,
-                plugin.name,
-                variant_name=variant,
-            )
-            plugin_def = self.discovery_service.find_definition(
-                plugin.type,
-                plugin.name,
-            )
-        else:
-            # Recursively look for the parent plugin definition and lock that
-            self.save(plugin.parent, overwrite=overwrite, exists_ok=True)
-            return
+        plugin_def = plugin.definition
+        path = self.projet.plugin_lock_path(
+            plugin_def.type,
+            plugin_def.name,
+            variant_name=variant,
+        )
 
         if path.exists() and not overwrite and not exists_ok:
             raise LockfileAlreadyExistsError(
@@ -102,6 +82,7 @@ class PluginLockService:
             plugin.name,
             plugin.namespace,
             plugin.type,
+            label=plugin.label,
         )
 
         with path.open("w") as lockfile: