+++ tests/fixtures/core.py
@@ -4,6 +4,7 @@ import logging
 import os
 import shutil
 from collections import namedtuple
+from copy import deepcopy
 from pathlib import Path
 
 import pytest
@@ -11,7 +12,6 @@ import yaml
 
 from meltano.core import bundle
 from meltano.core.behavior.canonical import Canonical
-from meltano.core.compiler.project_compiler import ProjectCompiler
 from meltano.core.config_service import ConfigService
 from meltano.core.elt_context import ELTContextBuilder
 from meltano.core.environment_service import EnvironmentService
@@ -79,6 +79,12 @@ def discovery():  # noqa: WPS213
                         },
                         {"name": "auth.username"},
                         {"name": "auth.password", "kind": "password"},
+                        {
+                            "name": "aliased",
+                            "kind": "string",
+                            "aliases": ["aliased_1", "aliased_2", "aliased_3"],
+                            "env_aliases": ["TAP_MOCK_ALIASED_ALIASED_1"],
+                        },
                     ],
                     "commands": {
                         "cmd": {
@@ -130,7 +136,12 @@ def discovery():  # noqa: WPS213
             "name": "target-mock",
             "namespace": "mock",
             "pip_url": "target-mock",
-            "settings": [{"name": "schema", "env": "MOCKED_SCHEMA"}],
+            "settings": [
+                {
+                    "name": "schema",
+                    "env": "MOCKED_SCHEMA",
+                }
+            ],
         }
     )
 
@@ -140,22 +151,6 @@ def discovery():  # noqa: WPS213
             "namespace": "tap_mock",
             "pip_url": "tap-mock-transform",
             "package_name": "dbt_mock",
-        }
-    )
-
-    discovery[PluginType.MODELS].append(
-        {
-            "name": "model-gitlab",
-            "namespace": "tap_gitlab",
-            "pip_url": "git+https://gitlab.com/meltano/model-gitlab.git",
-        }
-    )
-
-    discovery[PluginType.DASHBOARDS].append(
-        {
-            "name": "dashboard-google-analytics",
-            "namespace": "tap_google_analytics",
-            "pip_url": "git+https://gitlab.com/meltano/dashboard-google-analytics.git",
         }
     )
 
@@ -172,6 +167,14 @@ def discovery():  # noqa: WPS213
             "name": "transformer-mock",
             "namespace": "transformer_mock",
             "pip_url": "transformer-mock",
+            "requires": {
+                "files": [
+                    {
+                        "name": "files-transformer-mock",
+                        "variant": "meltano",
+                    },
+                ],
+            },
         }
     )
 
@@ -214,7 +217,13 @@ def discovery():  # noqa: WPS213
                     "executable": "mapper-mock-cmd",
                     "pip_url": "mapper-mock",
                     "package_name": "mapper-mock",
-                }
+                },
+                {
+                    "name": "alternative",
+                    "executable": "mapper-mock-alt",
+                    "pip_url": "mapper-mock-alt",
+                    "package_name": "mapper-mock-alt",
+                },
             ],
         }
     )
@@ -224,17 +233,12 @@ def discovery():  # noqa: WPS213
 
 @pytest.fixture(scope="class")
 def plugin_discovery_service(project, discovery):
-    return PluginDiscoveryService(project, discovery=discovery)
+    return PluginDiscoveryService(project, discovery=deepcopy(discovery))
 
 
 @pytest.fixture(scope="class")
 def locked_definition_service(project):
     return LockedDefinitionService(project)
-
-
-@pytest.fixture(scope="class")
-def project_compiler(project):
-    return ProjectCompiler(project)
 
 
 @pytest.fixture(scope="class")
@@ -279,43 +283,22 @@ def plugin_invoker_factory(
 
 
 @pytest.fixture(scope="class")
-def add_model(project, plugin_install_service, project_add_service):
-    models = [
-        "model-carbon-intensity",
-        "model-gitflix",
-        "model-salesforce",
-        "model-gitlab",
-    ]
-
-    for model in models:
-        plugin = project_add_service.add(PluginType.MODELS, model)
-        plugin_install_service.install_plugin(plugin)
-
-    yield
-
-    # clean-up
-    with project.meltano_update() as meltano:
-        meltano["plugins"]["models"] = [
-            model_def
-            for model_def in meltano["plugins"]["models"]
-            if model_def["name"] not in models
-        ]
-
-    for created_model in models:
-        shutil.rmtree(project.model_dir(created_model))
-
-
-@pytest.fixture(scope="class")
 def config_service(project):
     return ConfigService(project, use_cache=False)
 
 
 @pytest.fixture(scope="class")
-def project_plugins_service(project, config_service, plugin_discovery_service):
+def project_plugins_service(
+    project,
+    config_service,
+    plugin_discovery_service,
+    meltano_hub_service,
+):
     return ProjectPluginsService(
         project,
         config_service=config_service,
         discovery_service=plugin_discovery_service,
+        hub_service=meltano_hub_service,
         use_cache=False,
     )
 
@@ -400,6 +383,14 @@ def dbt(project_add_service):
 
 
 @pytest.fixture(scope="class")
+def transformer(project_add_service: ProjectAddService):
+    try:
+        return project_add_service.add(PluginType.TRANSFORMERS, "transformer-mock")
+    except PluginAlreadyAddedException as err:
+        return err.plugin
+
+
+@pytest.fixture(scope="class")
 def utility(project_add_service):
     try:
         return project_add_service.add(PluginType.UTILITIES, "utility-mock")
@@ -418,16 +409,28 @@ def task_sets_service(project):
 
 
 @pytest.fixture(scope="class")
-def schedule(project, tap, target, schedule_service):
+def elt_schedule(project, tap, target, schedule_service):
     try:
-        return schedule_service.add(
+        return schedule_service.add_elt(
             None,
-            "schedule-mock",
+            "elt-schedule-mock",
             extractor=tap.name,
             loader=target.name,
             transform="skip",
             interval="@once",
             start_date=datetime.datetime.now(),
+        )
+    except ScheduleAlreadyExistsError as err:
+        return err.schedule
+
+
+@pytest.fixture(scope="class")
+def job_schedule(project, tap, target, schedule_service):
+    try:
+        return schedule_service.add(
+            "job-schedule-mock",
+            "mock-job",
+            interval="@once",
         )
     except ScheduleAlreadyExistsError as err:
         return err.schedule
@@ -537,7 +540,7 @@ def mapper(project_add_service):
         return err.plugin
 
 
-def create_job_id(description: str, env: str = "dev") -> str:
+def create_state_id(description: str, env: str = "dev") -> str:
     return f"{env}:tap-{description}-to-target-{description}"
 
 
@@ -565,21 +568,21 @@ def payloads(num_params):
 
 
 @pytest.fixture
-def job_ids(num_params):
-    job_id_dict = {
-        "single_incomplete_job_id": create_job_id("single-incomplete"),
-        "single_complete_job_id": create_job_id("single-complete"),
-        "multiple_incompletes_job_id": create_job_id("multiple-incompletes"),
-        "multiple_completes_job_id": create_job_id("multiple-completes"),
-        "single_complete_then_multiple_incompletes_job_id": create_job_id(
+def state_ids(num_params):
+    state_id_dict = {
+        "single_incomplete_state_id": create_state_id("single-incomplete"),
+        "single_complete_state_id": create_state_id("single-complete"),
+        "multiple_incompletes_state_id": create_state_id("multiple-incompletes"),
+        "multiple_completes_state_id": create_state_id("multiple-completes"),
+        "single_complete_then_multiple_incompletes_state_id": create_state_id(
             "single-complete-then-multiple-incompletes"
         ),
-        "single_incomplete_then_multiple_completes_job_id": create_job_id(
+        "single_incomplete_then_multiple_completes_state_id": create_state_id(
             "single-incomplete-then-multiple-completes"
         ),
     }
-    job_ids = namedtuple("job_ids", job_id_dict)
-    return job_ids(**job_id_dict)
+    state_ids = namedtuple("state_ids", state_id_dict)
+    return state_ids(**state_id_dict)
 
 
 @pytest.fixture
@@ -605,63 +608,63 @@ def job_args():
 
 
 @pytest.fixture
-def job_ids_with_jobs(job_ids, job_args, payloads, mock_time):
+def state_ids_with_jobs(state_ids, job_args, payloads, mock_time):
     jobs = {
-        job_ids.single_incomplete_job_id: [
+        state_ids.single_incomplete_state_id: [
             Job(
-                job_id=job_ids.single_incomplete_job_id,
+                job_id=state_ids.single_incomplete_state_id,
                 **job_args.incomplete_job_args,
                 payload=payloads.mock_state_payloads[0],
             )
         ],
-        job_ids.single_complete_job_id: [
+        state_ids.single_complete_state_id: [
             Job(
-                job_id=job_ids.single_complete_job_id,
+                job_id=state_ids.single_complete_state_id,
                 payload=payloads.mock_state_payloads[0],
                 **job_args.complete_job_args,
             )
         ],
-        job_ids.multiple_incompletes_job_id: [
+        state_ids.multiple_incompletes_state_id: [
             Job(
-                job_id=job_ids.multiple_incompletes_job_id,
+                job_id=state_ids.multiple_incompletes_state_id,
                 **job_args.incomplete_job_args,
                 payload=payload,
             )
             for payload in payloads.mock_state_payloads
         ],
-        job_ids.multiple_completes_job_id: [
+        state_ids.multiple_completes_state_id: [
             Job(
-                job_id=job_ids.multiple_completes_job_id,
+                job_id=state_ids.multiple_completes_state_id,
                 payload=payload,
                 **job_args.complete_job_args,
             )
             for payload in payloads.mock_state_payloads
         ],
-        job_ids.single_complete_then_multiple_incompletes_job_id: [
+        state_ids.single_complete_then_multiple_incompletes_state_id: [
             Job(
-                job_id=job_ids.single_complete_then_multiple_incompletes_job_id,
+                job_id=state_ids.single_complete_then_multiple_incompletes_state_id,
                 payload=payloads.mock_state_payloads[0],
                 **job_args.complete_job_args,
             )
         ]
         + [
             Job(
-                job_id=job_ids.single_complete_then_multiple_incompletes_job_id,
+                job_id=state_ids.single_complete_then_multiple_incompletes_state_id,
                 payload=payload,
                 **job_args.incomplete_job_args,
             )
             for payload in payloads.mock_state_payloads[1:]
         ],
-        job_ids.single_incomplete_then_multiple_completes_job_id: [
+        state_ids.single_incomplete_then_multiple_completes_state_id: [
             Job(
-                job_id=job_ids.single_incomplete_then_multiple_completes_job_id,
+                job_id=state_ids.single_incomplete_then_multiple_completes_state_id,
                 payload=payloads.mock_state_payloads[0],
                 **job_args.incomplete_job_args,
             )
         ]
         + [
             Job(
-                job_id=job_ids.single_incomplete_then_multiple_completes_job_id,
+                job_id=state_ids.single_incomplete_then_multiple_completes_state_id,
                 payload=payload,
                 **job_args.complete_job_args,
             )
@@ -676,22 +679,24 @@ def job_ids_with_jobs(job_ids, job_args, payloads, mock_time):
 
 
 @pytest.fixture
-def jobs(job_ids_with_jobs):
-    return [job for job_list in job_ids_with_jobs.values() for job in job_list]
+def jobs(state_ids_with_jobs):
+    return [job for job_list in state_ids_with_jobs.values() for job in job_list]
 
 
 @pytest.fixture
-def job_ids_with_expected_states(job_ids, payloads, job_ids_with_jobs):  # noqa: WPS210
+def state_ids_with_expected_states(  # noqa: WPS210
+    state_ids, payloads, state_ids_with_jobs
+):
     final_state = {}
     for state in payloads.mock_state_payloads:
         merge(state, final_state)
     expectations = {
-        job_ids.single_complete_job_id: payloads.mock_state_payloads[0],
-        job_ids.single_incomplete_job_id: payloads.mock_empty_payload,
+        state_ids.single_complete_state_id: payloads.mock_state_payloads[0],
+        state_ids.single_incomplete_state_id: payloads.mock_empty_payload,
     }
 
-    for job_id, job_list in job_ids_with_jobs.items():
-        expectations[job_id] = {}
+    for state_id, job_list in state_ids_with_jobs.items():
+        expectations[state_id] = {}
 
         complete_jobs = []
         incomplete_jobs = []
@@ -712,15 +717,15 @@ def job_ids_with_expected_states(job_ids, payloads, job_ids_with_jobs):  # noqa:
         if incomplete_jobs:
             latest_incomplete_job = max(incomplete_jobs, key=lambda _job: _job.ended_at)
         if latest_complete_job:
-            expectations[job_id] = merge(
-                expectations[job_id], latest_complete_job.payload
+            expectations[state_id] = merge(
+                expectations[state_id], latest_complete_job.payload
             )
 
         for job in incomplete_jobs:
             if (not latest_complete_job) or (
                 job.ended_at > latest_complete_job.ended_at
             ):
-                expectations[job_id] = merge(expectations[job_id], job.payload)
+                expectations[state_id] = merge(expectations[state_id], job.payload)
         # Get all dummy jobs since latest non-dummy job.
         for job in dummy_jobs:
             if (
@@ -729,10 +734,10 @@ def job_ids_with_expected_states(job_ids, payloads, job_ids_with_jobs):  # noqa:
                 (not latest_incomplete_job)
                 or (job.ended_at > latest_incomplete_job.ended_at)
             ):
-                expectations[job_id] = merge(expectations[job_id], job.payload)
+                expectations[state_id] = merge(expectations[state_id], job.payload)
     return [
-        (test_job_id, expected_state)
-        for test_job_id, expected_state in expectations.items()
+        (test_state_id, expected_state)
+        for test_state_id, expected_state in expectations.items()
     ]