+++ tests/meltano/cli/test_add.py
@@ -6,9 +6,7 @@ import yaml
 
 from asserts import assert_cli_runner
 from meltano.cli import cli
-from meltano.cli.utils import CliError
-from meltano.core.m5o.dashboards_service import DashboardsService
-from meltano.core.m5o.reports_service import ReportsService
+from meltano.core.hub import MeltanoHubService
 from meltano.core.plugin import PluginRef, PluginType, Variant
 from meltano.core.plugin.error import PluginNotFoundError
 from meltano.core.plugin.project_plugin import ProjectPlugin
@@ -16,8 +14,16 @@ from meltano.core.plugin_install_service import PluginInstallReason
 
 
 class TestCliAdd:
+    @pytest.fixture(autouse=True)
+    def patch_hub(self, meltano_hub_service: MeltanoHubService):
+        with mock.patch(
+            "meltano.core.project_plugins_service.MeltanoHubService",
+            return_value=meltano_hub_service,
+        ):
+            yield
+
     @pytest.mark.parametrize(
-        "plugin_type,plugin_name,default_variant,related_plugin_refs",
+        "plugin_type,plugin_name,default_variant,required_plugin_refs",
         [
             (PluginType.EXTRACTORS, "tap-carbon-intensity", "meltano", []),
             (PluginType.LOADERS, "target-sqlite", "meltanolabs", []),
@@ -31,13 +37,19 @@ class TestCliAdd:
                 [PluginRef(PluginType.FILES, "airflow")],
             ),
         ],
+        ids=[
+            "single-extractor",
+            "single-loader",
+            "transform-and-related",
+            "orchestrator-and-required",
+        ],
     )
     def test_add(
         self,
         plugin_type,
         plugin_name,
         default_variant,
-        related_plugin_refs,
+        required_plugin_refs,
         project,
         cli_runner,
         project_plugins_service,
@@ -147,31 +159,6 @@ class TestCliAdd:
             "schema": "{{ env_var('DBT_SOURCE_SCHEMA') }}"
         }
 
-    def test_add_dashboard(self, project, cli_runner):
-        def install():
-            return cli_runner.invoke(
-                cli, ["add", "dashboard", "dashboard-google-analytics"]
-            )
-
-        res = install()
-        assert res.exit_code == 0
-
-        dashboards_service = DashboardsService(project)
-        dashboards_count = len(dashboards_service.get_dashboards())
-
-        assert dashboards_count > 0
-
-        reports_service = ReportsService(project)
-        reports_count = len(reports_service.get_reports())
-        assert reports_count > 0
-
-        # Verify that reinstalling doesn't duplicate dashboards and reports
-        res = install()
-        assert res.exit_code == 0
-
-        assert len(dashboards_service.get_dashboards()) == dashboards_count
-        assert len(reports_service.get_reports()) == reports_count
-
     def test_add_files_with_updates(
         self,
         project,
@@ -266,14 +253,21 @@ class TestCliAdd:
         ), mock.patch("meltano.cli.add.install_plugins") as install_plugin_mock:
             install_plugin_mock.return_value = True
             res = cli_runner.invoke(
-                cli, ["add", "extractor", "tap-mock", "--variant", "singer-io"]
+                cli,
+                [
+                    "add",
+                    "mapper",
+                    "mapper-mock",
+                    "--variant",
+                    "alternative",
+                ],
             )
             assert_cli_runner(res)
 
             plugin = project_plugins_service.find_plugin(
-                plugin_type=PluginType.EXTRACTORS, plugin_name="tap-mock"
+                plugin_type=PluginType.MAPPERS, plugin_name="mapper-mock"
             )
-            assert plugin.variant == "singer-io"
+            assert plugin.variant == "alternative"
 
     def test_add_inherited(
         self,
@@ -332,8 +326,7 @@ class TestCliAdd:
             )
             assert_cli_runner(res)
             assert (
-                "Inherit from:\ttap-mock, variant singer-io (deprecated)\n"
-                in res.stdout
+                "Inherit from:\ttap-mock, variant singer-io (default)\n" in res.stdout
             )
 
             inherited_variant = project_plugins_service.find_plugin(